requirements.txt
Django==2.0.7
django-filter==2.0.0
django-configurations
djangorestframework==3.8.2
Markdown==2.6.11
pytz==2018.5